# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Continuous Compliance Framework.",
        "title": "Continuous Compliance Framework API",
        "contact": {},
        "version": "1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/oscal/component-definitions": {
            "get": {
                "description": "Retrieves all component definitions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oscal"
                ],
                "summary": "List component definitions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GenericDataListResponse-oscal_List_responseComponentDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/oscal/component-definitions/{id}": {
            "get": {
                "description": "Retrieves a single component definition by its unique ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oscal"
                ],
                "summary": "Get a component definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Component Definition ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GenericDataResponse-oscal_Get_responseComponentDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "handler.GenericDataListResponse-oscal_List_responseComponentDefinition": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Items from the list response",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscal.List.responseComponentDefinition"
                    }
                }
            }
        },
...
        "handler.GenericDataResponse-oscal_Get_responseComponentDefinition": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Items from the list response",
                    "allOf": [
                        {
                            "$ref": "#/definitions/oscal.Get.responseComponentDefinition"
                        }
                    ]
                }
            }
        },
...
        "oscal.Get.responseComponentDefinition": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/oscalTypes_1_1_3.Metadata"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "oscal.List.responseComponentDefinition": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/oscalTypes_1_1_3.Metadata"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.Action": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Link"
                    }
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Property"
                    }
                },
                "remarks": {
                    "type": "string"
                },
                "responsible-parties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.ResponsibleParty"
                    }
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.Address": {
            "type": "object",
            "properties": {
                "addr-lines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "postal-code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "oscalTypes_1_1_3.DocumentId": {
            "type": "object",
            "properties": {
                "identifier": {
                    "type": "string"
                },
                "scheme": {
                    "type": "string"
                }
            }
        },
...
        "oscalTypes_1_1_3.Link": {
            "type": "object",
            "properties": {
                "href": {
                    "type": "string"
                },
                "media-type": {
                    "type": "string"
                },
                "rel": {
                    "type": "string"
                },
                "resource-fragment": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.Location": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/oscalTypes_1_1_3.Address"
                },
                "email-addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Link"
                    }
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Property"
                    }
                },
                "remarks": {
                    "type": "string"
                },
                "telephone-numbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.TelephoneNumber"
                    }
                },
                "title": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.Metadata": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Action"
                    }
                },
                "document-ids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.DocumentId"
                    }
                },
                "last-modified": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Link"
                    }
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Location"
                    }
                },
                "oscal-version": {
                    "type": "string"
                },
                "parties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Party"
                    }
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Property"
                    }
                },
                "published": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "responsible-parties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.ResponsibleParty"
                    }
                },
                "revisions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.RevisionHistoryEntry"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Role"
                    }
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "oscalTypes_1_1_3.Party": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Address"
                    }
                },
                "email-addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "external-ids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.PartyExternalIdentifier"
                    }
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Link"
                    }
                },
                "location-uuids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "member-of-organizations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Property"
                    }
                },
                "remarks": {
                    "type": "string"
                },
                "short-name": {
                    "type": "string"
                },
                "telephone-numbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.TelephoneNumber"
                    }
                },
                "type": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.PartyExternalIdentifier": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "scheme": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.Property": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ns": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "oscalTypes_1_1_3.ResponsibleParty": {
            "type": "object",
            "properties": {
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Link"
                    }
                },
                "party-uuids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Property"
                    }
                },
                "remarks": {
                    "type": "string"
                },
                "role-id": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.RevisionHistoryEntry": {
            "type": "object",
            "properties": {
                "last-modified": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Link"
                    }
                },
                "oscal-version": {
                    "type": "string"
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Property"
                    }
                },
                "published": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "oscalTypes_1_1_3.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Link"
                    }
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oscalTypes_1_1_3.Property"
                    }
                },
                "remarks": {
                    "type": "string"
                },
                "short-name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "oscalTypes_1_1_3.TelephoneNumber": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
...
}